# Speed-up `QubitMapper.mode_based_mapping`