<!--
⚠️ If you do not respect this template, your pull request will be closed.
⚠️ Your pull request title should be short detailed and understandable for all.
⚠️ Also, please add it in the CHANGELOG file under Unreleased section.
⚠️ If your pull request fixes an open issue, please link to the issue.

✅ I have added the tests to cover my changes.
✅ I have updated the documentation accordingly.
✅ I have read the CONTRIBUTING document.
-->

### Summary

This PR improves performance of JW transformation by
- computing `SparsePauliOp` of "+", "-", "E", "N" beforehand
- reducing the overhead of `SparsePauliOp._add`

If we introduce a specialized function to compute the sum of `SparsePauliOp` (e.g., `SparsePauliOp.sum(list[SparsePauliOp]) -> SparsePauliOp`, we can further improve the performance around 2x. But, we need discussion to introduce a new public API to Terra.

Update: `SpasePauliOp.sum` is merged https://github.com/Qiskit/qiskit-terra/pull/7202. So, I replaced the workaround of sum with `SparsePauliOp.sum`.

### Details and comments
I also prepared separate PRs to improve the performance of JW as follows:
- https://github.com/Qiskit/qiskit-terra/pull/7122
- https://github.com/Qiskit/qiskit-terra/pull/7126
- https://github.com/Qiskit/qiskit-terra/pull/7138

I notice that `SparsePauliOp.simplify` has a critical performance issue, which is addressed at  https://github.com/Qiskit/qiskit-terra/pull/7122, so, this PR with terra main still takes a long time.

Microbenchmark

| Nature branch  | Terra branch | Time |
|----------------|--------------|------|
| main | main | 61.88646222 sec  |
| this PR | main | 53.311054609 sec |
| main | stable | 12.960038270999998 sec |
| this PR | stable | 10.459337683 sec |
| main | https://github.com/Qiskit/qiskit-terra/pull/7122 | 12.132281858 sec  |
| this PR | https://github.com/Qiskit/qiskit-terra/pull/7122 | 4.158444311 sec  |
| this PR | 7122 + 7126 + 7138 | 1.7728687539999999 sec |

Update: the final version with SparsePauliOp.sum takes 1.63 sec.


```python
import random
from timeit import timeit

from qiskit_nature.converters.second_quantization import QubitConverter
from qiskit_nature.mappers.second_quantization import JordanWignerMapper
from qiskit_nature.operators.second_quantization import FermionicOp

random.seed(123)
k = 10
n = 100

op = sum(
    (FermionicOp(''.join(random.choices('+-ENI', k=k)), display_format="dense") * random.random())
    for _ in range(n))
mapper = JordanWignerMapper()
qubit_conv = QubitConverter(mapper)
print(f"{timeit(lambda: qubit_conv.convert(op), number=1)} sec")
```