@t-imamichi 2021-11-03
Because https://github.com/Qiskit/qiskit-terra/pull/7202 is merged, I replace the workaround of `sum` with `SpasePauliOp.sum`. It gets even faster.