@mrossinek 2021-11-05
@woodsp-ibm Offline you mentioned that you might have some more comments on this PR. Is that still the case or can this be merged?